mainrepo/lib/bi_formula/pyproject.toml

[tool.poetry]
name = "yandex-bi-formula"
version = "7.47.0"
description = ""
authors = ["Grigory Statsenko <gstatsenko@yandex-team.ru>"]
packages = [{include = "bi_formula"}]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
antlr4-python3-runtime = ">=4.11.1"
attrs = ">=22.2.0"
ciso8601 = ">=2.2.0"
graphviz = ">=0.17"
jinja2 = ">=3.1.2"
python = ">=3.10, <3.12"
pytz = ">=2022.7.1"
sqlalchemy = ">=1.4.46, <2.0"
sqlalchemy-gsheets = {path = "../../../lib/bi_sqlalchemy_gsheets"}
yandex-bi-sqlalchemy-bitrix = {path = "../../../lib/bi_sqlalchemy_bitrix"}
yandex-bi-db-testing = {path = "../bi_db_testing"}
yandex-bi-utils = {path = "../bi_utils"}
dynamic-enum = {path = "../dynamic_enum"}

[tool.poetry.group]
[tool.poetry.group.tests.dependencies]
pytest = ">=7.2.2"

[tool.poetry.group.tests_unit.dependencies]
mysqlclient = ">=1.4.6"
oracledb = ">=1.3.1"
psycopg2-binary = ">=2.9.6"
pyodbc = ">=4.0.35"
pytest = ">=7.2.2"
python-dateutil = ">=2.8.2"
sqlalchemy-metrika-api = {path = "../../../lib/sqlalchemy_metrika_api"}
statinfra-clickhouse-sqlalchemy = {path = "../../lib/clickhouse-sqlalchemy"}
yandex-bi-connector-clickhouse = {path = "../../lib/bi_connector_clickhouse"}
yandex-bi-connector-postgresql = {path = "../../lib/bi_connector_postgresql"}
yandex-bi-sqlalchemy-yq = {path = "../../../lib/bi_sqlalchemy_yq"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = ["bi_formula_tests/unit"]

[tool.ruff]
line-length = 140
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I"
]

[tool.black]
line-length = 140
target-version = ['py310']

[tool.poetry.scripts]
bi-formula-cli = "bi_formula.scripts.formula_cli:main"
